+++ relay-conventions/src/lib.rs
@@ -54,6 +54,7 @@
 //! ### I want to reference an attribute in Relay but it's not defined in `sentry-conventions`, what should I do?
 //! **Always** define it in `sentry-conventions` before using it in Relay. This makes sure we have proper
 pub mod consts {
+    //! Attribute constant definitions.
     #![allow(rustdoc::bare_urls)]
     #![allow(non_upper_case_globals)]
     include!(concat!(env!("OUT_DIR"), "/attribute_consts.rs"));
@@ -72,7 +73,14 @@ pub mod consts {
     pub use self::not_yet_defined::*;
 }
 
+pub mod interpolate {
+    //! Functions for interpolating attribute keys with placeholders.
+    #![allow(non_snake_case)]
+    include!(concat!(env!("OUT_DIR"), "/interpolation_fns.rs"));
+}
+
 include!(concat!(env!("OUT_DIR"), "/attribute_map.rs"));
+include!(concat!(env!("OUT_DIR"), "/canonical_fn.rs"));
 include!(concat!(env!("OUT_DIR"), "/name_fn.rs"));
 
 /// Whether an attribute should be PII-strippable/should be subject to datascrubbers